LibWeb: Carry gradient interpolation as a Gfx type
GradientData held the CSS style-value representation of the color
interpolation method, and the display list recorder converted it to
Gfx::GradientInterpolationMethod while building each gradient command.
The recorder is a byte-appending seam, not a place for CSS-to-Gfx
translation, and the commands already carry the Gfx type. Convert at
gradient resolution instead, so GradientData stores the Gfx type and
the recorder copies it through unchanged.
