+++ Libraries/LibWeb/Painting/GradientPainting.cpp
@@ -21,6 +21,85 @@
 
 namespace Web::Painting {
 
+static Gfx::RectangularColorSpace to_gfx_rectangular_color_space(CSS::RectangularColorSpace color_space)
+{
+    switch (color_space) {
+    case CSS::RectangularColorSpace::Srgb:
+        return Gfx::RectangularColorSpace::Srgb;
+    case CSS::RectangularColorSpace::SrgbLinear:
+        return Gfx::RectangularColorSpace::SrgbLinear;
+    case CSS::RectangularColorSpace::DisplayP3:
+        return Gfx::RectangularColorSpace::DisplayP3;
+    case CSS::RectangularColorSpace::DisplayP3Linear:
+        return Gfx::RectangularColorSpace::DisplayP3Linear;
+    case CSS::RectangularColorSpace::A98Rgb:
+        return Gfx::RectangularColorSpace::A98Rgb;
+    case CSS::RectangularColorSpace::ProphotoRgb:
+        return Gfx::RectangularColorSpace::ProphotoRgb;
+    case CSS::RectangularColorSpace::Rec2020:
+        return Gfx::RectangularColorSpace::Rec2020;
+    case CSS::RectangularColorSpace::Lab:
+        return Gfx::RectangularColorSpace::Lab;
+    case CSS::RectangularColorSpace::Oklab:
+        return Gfx::RectangularColorSpace::Oklab;
+    case CSS::RectangularColorSpace::Xyz:
+        return Gfx::RectangularColorSpace::Xyz;
+    case CSS::RectangularColorSpace::XyzD50:
+        return Gfx::RectangularColorSpace::XyzD50;
+    case CSS::RectangularColorSpace::XyzD65:
+        return Gfx::RectangularColorSpace::XyzD65;
+    }
+    VERIFY_NOT_REACHED();
+}
+
+static Gfx::PolarColorSpace to_gfx_polar_color_space(CSS::PolarColorSpace color_space)
+{
+    switch (color_space) {
+    case CSS::PolarColorSpace::Hsl:
+        return Gfx::PolarColorSpace::Hsl;
+    case CSS::PolarColorSpace::Hwb:
+        return Gfx::PolarColorSpace::Hwb;
+    case CSS::PolarColorSpace::Lch:
+        return Gfx::PolarColorSpace::Lch;
+    case CSS::PolarColorSpace::Oklch:
+        return Gfx::PolarColorSpace::Oklch;
+    }
+    VERIFY_NOT_REACHED();
+}
+
+static Gfx::HueInterpolationMethod to_gfx_hue_interpolation_method(CSS::HueInterpolationMethod hue_interpolation_method)
+{
+    switch (hue_interpolation_method) {
+    case CSS::HueInterpolationMethod::Shorter:
+        return Gfx::HueInterpolationMethod::Shorter;
+    case CSS::HueInterpolationMethod::Longer:
+        return Gfx::HueInterpolationMethod::Longer;
+    case CSS::HueInterpolationMethod::Increasing:
+        return Gfx::HueInterpolationMethod::Increasing;
+    case CSS::HueInterpolationMethod::Decreasing:
+        return Gfx::HueInterpolationMethod::Decreasing;
+    }
+    VERIFY_NOT_REACHED();
+}
+
+static Gfx::GradientInterpolationMethod to_display_list_color_interpolation_method(
+    CSS::ColorInterpolationMethodStyleValue::ColorInterpolationMethod const& interpolation_method)
+{
+    Gfx::GradientInterpolationMethod result;
+    interpolation_method.visit(
+        [&](CSS::RectangularColorSpace color_space) {
+            result.type = Gfx::GradientInterpolationMethod::Type::Rectangular;
+            result.rectangular_color_space = to_gfx_rectangular_color_space(color_space);
+        },
+        [&](CSS::ColorInterpolationMethodStyleValue::PolarColorInterpolationMethod const& color_space) {
+            result.type = Gfx::GradientInterpolationMethod::Type::Polar;
+            result.polar_color_space = to_gfx_polar_color_space(color_space.color_space);
+            result.hue_interpolation_method = to_gfx_hue_interpolation_method(color_space.hue_interpolation_method);
+        });
+    return result;
+}
+
+
 using ColorStopList = Vector<Gfx::ColorStop, 4>;
 
 struct ResolvedColorStopData {
@@ -330,7 +409,7 @@ LinearGradientData resolve_linear_gradient_data(Layout::NodeWithStyle const& nod
     auto color_stop_data = to_color_stop_data(resolved_color_stops.list, resolved_color_stops.repeating, [&](float position) {
         return (position - first_stop_position) / repeat_length;
     });
-    return { gradient_angle, first_stop_position, repeat_length, move(color_stop_data), linear_gradient.interpolation_method() };
+    return { gradient_angle, first_stop_position, repeat_length, move(color_stop_data), to_display_list_color_interpolation_method(linear_gradient.interpolation_method()) };
 }
 
 ConicGradientData resolve_conic_gradient_data(Layout::NodeWithStyle const& node, CSS::ConicGradientStyleValue const& conic_gradient)
@@ -352,7 +431,7 @@ ConicGradientData resolve_conic_gradient_data(Layout::NodeWithStyle const& node,
     auto color_stop_data = to_color_stop_data(resolved_color_stops.list, resolved_color_stops.repeating, [](float position) {
         return position;
     });
-    return { conic_gradient.angle_degrees(), move(color_stop_data), conic_gradient.interpolation_method() };
+    return { conic_gradient.angle_degrees(), move(color_stop_data), to_display_list_color_interpolation_method(conic_gradient.interpolation_method()) };
 }
 
 RadialGradientData resolve_radial_gradient_data(Layout::NodeWithStyle const& node, CSSPixelSize gradient_size, CSS::RadialGradientStyleValue const& radial_gradient)
@@ -374,7 +453,7 @@ RadialGradientData resolve_radial_gradient_data(Layout::NodeWithStyle const& nod
     auto color_stop_data = to_color_stop_data(resolved_color_stops.list, resolved_color_stops.repeating, [](float position) {
         return position;
     });
-    return { move(color_stop_data), radial_gradient.interpolation_method() };
+    return { move(color_stop_data), to_display_list_color_interpolation_method(radial_gradient.interpolation_method()) };
 }
 
 }
